+++ src/components/CategoryGrid.astro
@@ -5,7 +5,7 @@ import {
   useTranslatedPath,
 } from '../i18n/utils';
 import { defaultLang } from '../i18n/ui';
-import { getCategoryConfigs } from '../utils/newCategoryConfig';
+import { getCategoryConfigs } from '../utils/categoryConfig';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
