+++ src/pages/[category]/index.astro
@@ -9,7 +9,7 @@ import {
   categoryList as categories,
   getCategoryConfigs,
   type CategoryKey,
-} from '../../utils/newCategoryConfig';
+} from '../../utils/categoryConfig';
 
 export async function getStaticPaths() {
   const categoryMapping: Record<string, string> = {
