+++ src/pages/en/[category]/[slug].astro
@@ -10,7 +10,7 @@ import {
   categoryList as categories,
   getCategoryConfigs,
   type CategoryKey,
-} from '../../../utils/newCategoryConfig';
+} from '../../../utils/categoryConfig';
 
 let _gitCache: Map<
   string,
