+++ src/pages/en/[category]/index.astro
@@ -10,7 +10,7 @@ import {
   categoryList as categories,
   getCategoryConfigs,
   type CategoryKey,
-} from '../../../utils/newCategoryConfig';
+} from '../../../utils/categoryConfig';
 
 export async function getStaticPaths() {
   const categoryMapping: Record<string, string> = {
